lab1/src/main/java/ar/uba/fi/compiladores/parte3/ManualLexer.java
package ar.uba.fi.compiladores.parte3;

import java.util.List;

/**
 * Clase que implementa un algoritmo de lexeo a partir de un lenguage.
 */
public class ManualLexer<S,T>{
    private LanguageAutomata<S,T> language;
    public ManualLexer(LanguageAutomata<S,T> language){
        this.language=language;
    }
    /**
     * extracts a token from a string and returns the next state of the LanguageAutomata and the length of the extracted token.
     * @param program
     * @return
     */
    public SingleTokenExtraction<S> extractToken(String program){
        this.language.eofTokenType();
        return null;
    }
    public List<Token<T>> lex(String program){
        return null;
    }

}